+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/actuators/actuator_base.py
@@ -18,36 +18,60 @@ if TYPE_CHECKING:
 
 
 class ActuatorBase(ABC):
-    """Base class for applying actuator models over a collection of actuated joints in an articulation.
+    """Base class for actuator models over a collection of actuated joints in an articulation.
 
-    The default actuator for applying the same actuator model over a collection of actuated joints in
-    an articulation.
+    Actuator models augment the simulated articulation joints with an external drive dynamics model.
+    The model is used to convert the user-provided joint commands (positions, velocities and efforts)
+    into the desired joint positions, velocities and efforts that are applied to the simulated articulation.
 
-    The joint names are specified in the configuration through a list of regular expressions. The regular
-    expressions are matched against the joint names in the articulation. The first match is used to determine
-    the joint indices in the articulation.
+    The base class provides the interface for the actuator models. It is responsible for parsing the
+    actuator parameters from the configuration and storing them as buffers. It also provides the
+    interface for resetting the actuator state and computing the desired joint commands for the simulation.
 
-    In the default actuator, no constraints or formatting is performed over the input actions. Thus, the
-    input actions are directly used to compute the joint actions in the :meth:`compute`.
+    For each actuator model, a corresponding configuration class is provided. The configuration class
+    is used to parse the actuator parameters from the configuration. It also specifies the joint names
+    for which the actuator model is applied. These names can be specified as regular expressions, which
+    are matched against the joint names in the articulation.
+
+    To see how the class is used, check the :class:`omni.isaac.orbit.assets.Articulation` class.
     """
 
     computed_effort: torch.Tensor
     """The computed effort for the actuator group. Shape is ``(num_envs, num_joints)``."""
     applied_effort: torch.Tensor
     """The applied effort for the actuator group. Shape is ``(num_envs, num_joints)``."""
-    effort_limit: float = torch.inf
+    effort_limit: torch.Tensor
     """The effort limit for the actuator group. Shape is ``(num_envs, num_joints)``."""
-    velocity_limit: float = torch.inf
+    velocity_limit: torch.Tensor
     """The velocity limit for the actuator group. Shape is ``(num_envs, num_joints)``."""
     stiffness: torch.Tensor
     """The stiffness (P gain) of the PD controller. Shape is ``(num_envs, num_joints)``."""
     damping: torch.Tensor
     """The damping (D gain) of the PD controller. Shape is ``(num_envs, num_joints)``."""
+    armature: torch.Tensor
+    """The armature of the actuator joints. Shape is ``(num_envs, num_joints)``."""
+    friction: torch.Tensor
+    """The joint friction of the actuator joints. Shape is ``(num_envs, num_joints)``."""
 
     def __init__(
-        self, cfg: ActuatorBaseCfg, joint_names: list[str], joint_ids: Sequence[int], num_envs: int, device: str
+        self,
+        cfg: ActuatorBaseCfg,
+        joint_names: list[str],
+        joint_ids: Sequence[int],
+        num_envs: int,
+        device: str,
+        stiffness: torch.Tensor | float = 0.0,
+        damping: torch.Tensor | float = 0.0,
+        armature: torch.Tensor | float = 0.0,
+        friction: torch.Tensor | float = 0.0,
+        effort_limit: torch.Tensor | float = torch.inf,
+        velocity_limit: torch.Tensor | float = torch.inf,
     ):
         """Initialize the actuator.
+
+        Note:
+            The actuator parameters are parsed from the configuration and stored as buffers. If the parameters
+            are not specified in the configuration, then the default values provided in the arguments are used.
 
         Args:
             cfg: The configuration of the actuator model.
@@ -56,6 +80,18 @@ class ActuatorBase(ABC):
                 the joints in the articulation are part of the group.
             num_envs: Number of articulations in the view.
             device: Device used for processing.
+            stiffness: The default joint stiffness (P gain). Defaults to 0.0.
+                If a tensor, then the shape is ``(num_envs, num_joints)``.
+            damping: The default joint damping (D gain). Defaults to 0.0.
+                If a tensor, then the shape is ``(num_envs, num_joints)``.
+            armature: The default joint armature. Defaults to 0.0.
+                If a tensor, then the shape is ``(num_envs, num_joints)``.
+            friction: The default joint friction. Defaults to 0.0.
+                If a tensor, then the shape is ``(num_envs, num_joints)``.
+            effort_limit: The default effort limit. Defaults to infinity.
+                If a tensor, then the shape is ``(num_envs, num_joints)``.
+            velocity_limit: The default velocity limit. Defaults to infinity.
+                If a tensor, then the shape is ``(num_envs, num_joints)``.
         """
         # save parameters
         self.cfg = cfg
@@ -64,36 +100,20 @@ class ActuatorBase(ABC):
         self._joint_names = joint_names
         self._joint_indices = joint_ids
 
+        # parse joint stiffness and damping
+        self.stiffness = self._parse_joint_parameter(self.cfg.stiffness, stiffness)
+        self.damping = self._parse_joint_parameter(self.cfg.damping, damping)
+        # parse joint armature and friction
+        self.armature = self._parse_joint_parameter(self.cfg.armature, armature)
+        self.friction = self._parse_joint_parameter(self.cfg.friction, friction)
+        # parse joint limits
+        # note: for velocity limits, we don't have USD parameter, so default is infinity
+        self.effort_limit = self._parse_joint_parameter(self.cfg.effort_limit, effort_limit)
+        self.velocity_limit = self._parse_joint_parameter(self.cfg.velocity_limit, velocity_limit)
+
         # create commands buffers for allocation
         self.computed_effort = torch.zeros(self._num_envs, self.num_joints, device=self._device)
         self.applied_effort = torch.zeros_like(self.computed_effort)
-        self.stiffness = torch.zeros_like(self.computed_effort)
-        self.damping = torch.zeros_like(self.computed_effort)
-
-        # parse joint limits
-        if self.cfg.effort_limit is not None:
-            self.effort_limit = self.cfg.effort_limit
-        if self.cfg.velocity_limit is not None:
-            self.velocity_limit = self.cfg.velocity_limit
-        # parse joint stiffness and damping
-        # -- stiffness
-        if self.cfg.stiffness is not None:
-            if isinstance(self.cfg.stiffness, (float, int)):
-                # if float, then use the same value for all joints
-                self.stiffness[:] = float(self.cfg.stiffness)
-            else:
-                # if dict, then parse the regular expression
-                indices, _, values = string_utils.resolve_matching_names_values(self.cfg.stiffness, self.joint_names)
-                self.stiffness[:, indices] = torch.tensor(values, device=self._device)
-        # -- damping
-        if self.cfg.damping is not None:
-            if isinstance(self.cfg.damping, (float, int)):
-                # if float, then use the same value for all joints
-                self.damping[:] = float(self.cfg.damping)
-            else:
-                # if dict, then parse the regular expression
-                indices, _, values = string_utils.resolve_matching_names_values(self.cfg.damping, self.joint_names)
-                self.damping[:, indices] = torch.tensor(values, device=self._device)
 
     def __str__(self) -> str:
         """Returns: A string representation of the actuator group."""
@@ -164,3 +184,52 @@ class ActuatorBase(ABC):
             The computed desired joint positions, joint velocities and joint efforts.
         """
         raise NotImplementedError
+
+    """
+    Helper functions.
+    """
+
+    def _parse_joint_parameter(
+        self, cfg_value: float | dict[str, float] | None, default_value: float | torch.Tensor | None
+    ) -> torch.Tensor:
+        """Parse the joint parameter from the configuration.
+
+        Args:
+            cfg_value: The parameter value from the configuration. If None, then use the default value.
+            default_value: The default value to use if the parameter is None. If it is also None,
+                then an error is raised.
+
+        Returns:
+            The parsed parameter value.
+
+        Raises:
+            TypeError: If the parameter value is not of the expected type.
+            TypeError: If the default value is not of the expected type.
+            ValueError: If the parameter value is None and no default value is provided.
+        """
+        # create parameter buffer
+        param = torch.zeros(self._num_envs, self.num_joints, device=self._device)
+        # parse the parameter
+        if cfg_value is not None:
+            if isinstance(cfg_value, (float, int)):
+                # if float, then use the same value for all joints
+                param[:] = float(cfg_value)
+            elif isinstance(cfg_value, dict):
+                # if dict, then parse the regular expression
+                indices, _, values = string_utils.resolve_matching_names_values(param, self.joint_names)
+                param[:, indices] = torch.tensor(values, device=self._device)
+            else:
+                raise TypeError(f"Invalid type for parameter value: {type(cfg_value)}. Expected float or dict.")
+        elif default_value is not None:
+            if isinstance(default_value, (float, int)):
+                # if float, then use the same value for all joints
+                param[:] = float(default_value)
+            elif isinstance(default_value, torch.Tensor):
+                # if tensor, then use the same tensor for all joints
+                param[:] = default_value
+            else:
+                raise TypeError(f"Invalid type for default value: {type(default_value)}. Expected float or Tensor.")
+        else:
+            raise ValueError("The parameter value is None and no default value is provided.")
+
+        return param